Hydrates saved data with classes
